chore: use u64 to represent most integer newtypes in rust, and numeric in pgsql (#313)
* add fields returned from substrate

* add nonce

* add nonce to block

* fix constraint

* comments

* lint

* change default hashes

* fix dumb naming

* fmt

* Remove nonsensical comment

* rename block_nonce to miner_nonce

* fmt

* change Index internal type to u64

* change nonce to u64

* change chainid to u64

* change size to u64

* change gas to u64

* fix bugs

* make everything numeric

* cleanup

* fix

* lint

* sqlx fix

* fix queries

* fix sqlx conversions

* fmt

* remove test file

* fix sqlx conversions

* fmt

* sqlx prepare

* use U64 instead of u64

* change chainid to U64

* lint

+++ src/eth/block_miner.rs
@@ -72,7 +72,7 @@ impl BlockMiner {
         // mine transactions and logs
         let mut log_index = Index::ZERO;
         for (tx_idx, (input, execution)) in transactions.into_iter().enumerate() {
-            let transaction_index = Index::new(tx_idx as u16);
+            let transaction_index = Index::new(tx_idx as u64);
             // mine logs
             let mut mined_logs: Vec<LogMined> = Vec::with_capacity(execution.logs.len());
             for mined_log in execution.logs.clone() {
